+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project *loosely tries* to adhere to [Semantic Versioning](http://semver.org/), even before v1.0.
 
+## [0.8.16]
+- add minimal deploy iam policy docs
+- harden deploy, auto delete minimal stack in rollback_completed
+- Merge pull request #38 from tongueroo/harden-deploy-rollbacks
+
 ## [0.8.15]
 - fix route resources macro for api mode pull request #35
 - remove pg dependency from jets and add as part of project Gemfile pull request #36